requirements.txt
google-genai>=1.0.0
groq>=0.11.0
python-dotenv>=1.0.0
tenacity>=8.2.0
pandas>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
numpy>=1.24.0
